+++ src/sound/snd_pas16.c
@@ -79,7 +79,7 @@
  *		FF88 - board model
  *		  3 = PAS16
  *
- * Version:	@(#)snd_pas16.c	1.0.2	2018/03/15
+ * Version:	@(#)snd_pas16.c	1.0.3	2018/03/28
  *
  * Authors:	Fred N. van Kempen, <decwiz@yahoo.com>
  *		Miran Grca, <mgrca8@gmail.com>
@@ -443,9 +443,9 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                 {
                         if (!(val & 0x20))
                         {
-                                if (val & 2) pas16->pit.rl[0] = pas16->pit.c[0] / (PITCONST * (1 << TIMER_SHIFT));
-                                if (val & 4) pas16->pit.rl[1] = pas16->pit.c[1];
-                                if (val & 8) pas16->pit.rl[2] = pas16->pit.c[2];
+                                if (val & 2) pas16->pit.rl[0] = (int16_t) (pas16->pit.c[0] / (PITCONST * (1 << TIMER_SHIFT)));
+                                if (val & 4) pas16->pit.rl[1] = (int16_t) (pas16->pit.c[1]);
+                                if (val & 8) pas16->pit.rl[2] = (int16_t) (pas16->pit.c[2]);
                         }
                         return;
                 }
@@ -458,9 +458,9 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                 }
                 if (!(pas16->pit.ctrl & 0x30))
                 {
-                        pas16->pit.rl[t] = pas16->pit.c[t];
+                        pas16->pit.rl[t] = (int16_t)pas16->pit.c[t];
                         if (!t)
-                                pas16->pit.rl[t] /= (PITCONST * (1 << TIMER_SHIFT));
+                                pas16->pit.rl[t] /= (int16_t) ((PITCONST * (1 << TIMER_SHIFT)));
                         if (pas16->pit.c[t] < 0) 
                                 pas16->pit.rl[t] = 0;
                         pas16->pit.ctrl |= 0x30;
@@ -476,9 +476,9 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                         if (!pas16->pit.rm[t])
                         {
                                 pas16->pit.rm[t] = 3;
-                                pas16->pit.rl[t] = pit.c[t];
+                                pas16->pit.rl[t] = (int16_t)pit.c[t];
                                 if (!t)
-                                        pas16->pit.rl[t] /= (PITCONST * (1 << TIMER_SHIFT));
+                                        pas16->pit.rl[t] /= (int16_t) ((PITCONST * (1 << TIMER_SHIFT)));
                         }
                         pas16->pit.rereadlatch[t] = 1;
                 }
@@ -494,7 +494,7 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                         pas16->pit.thit[t] = 0;
                         pas16->pit.c[t] = pas16->pit.l[t];
                         if (!t)
-                                pas16->pit.c[t] *= PITCONST * (1 << TIMER_SHIFT);
+                                pas16->pit.c[t] *= (int64_t) (PITCONST * (1 << TIMER_SHIFT));
                         pas16->pit.enable[t] = 1;
                         break;
                         case 2:
@@ -502,7 +502,7 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                         pas16->pit.thit[t] = 0;
                         pas16->pit.c[t] = pas16->pit.l[t];
                         if (!t)
-                                pas16->pit.c[t] *= PITCONST * (1 << TIMER_SHIFT);
+                                pas16->pit.c[t] *= (int64_t) (PITCONST * (1 << TIMER_SHIFT));
                         pas16->pit.enable[t] = 1;
                         break;
                         case 0:
@@ -510,7 +510,7 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                         pas16->pit.l[t] |= (val << 8);
                         pas16->pit.c[t] = pas16->pit.l[t];
                         if (!t)
-                                pas16->pit.c[t] *= PITCONST * (1 << TIMER_SHIFT);
+                                pas16->pit.c[t] *= (int64_t) (PITCONST * (1 << TIMER_SHIFT));
                         pas16->pit.thit[t] = 0;
                         pas16->pit.wm[t] = 3;
                         pas16->pit.enable[t] = 1;
@@ -526,7 +526,7 @@ static void pas16_pit_out(uint16_t port, uint8_t val, void *p)
                         pas16->pit.l[t] |= 0x10000;
                         pas16->pit.c[t] = pas16->pit.l[t];
                         if (!t)
-                                pas16->pit.c[t] *= PITCONST * (1 << TIMER_SHIFT);
+                                pas16->pit.c[t] *= (int64_t) (PITCONST * (1 << TIMER_SHIFT));
                 }
                 break;
         }
@@ -545,13 +545,13 @@ static uint8_t pas16_pit_in(uint16_t port, void *p)
                         pas16->pit.rereadlatch[t] = 0;
                         if (!t)
                         {
-                                pas16->pit.rl[t] = pas16->pit.c[t] / (PITCONST * (1 << TIMER_SHIFT));
+                                pas16->pit.rl[t] = (uint16_t) (pas16->pit.c[t] / (PITCONST * (1 << TIMER_SHIFT)));
                                 if ((pas16->pit.c[t] / (PITCONST * (1 << TIMER_SHIFT))) > 65536) 
                                         pas16->pit.rl[t] = 0xFFFF;
                         }
                         else
                         {
-                                pas16->pit.rl[t] = pas16->pit.c[t];
+                                pas16->pit.rl[t] = (uint16_t)pas16->pit.c[t];
                                 if (pas16->pit.c[t] > 65536) 
                                         pas16->pit.rl[t] = 0xFFFF;
                         }
@@ -598,9 +598,9 @@ static void pas16_pcm_poll(void *p)
         if (pas16->pit.m[0] & 2)
         {
                 if (pas16->pit.l[0]) 
-                        pas16->pit.c[0] += (pas16->pit.l[0] * PITCONST * (1 << TIMER_SHIFT));
+                        pas16->pit.c[0] += (int64_t) ((pas16->pit.l[0] * PITCONST * (1 << TIMER_SHIFT)));
                 else                
-                        pas16->pit.c[0] += (0x10000 * PITCONST * (1 << TIMER_SHIFT));
+                        pas16->pit.c[0] += (int64_t) ((0x10000 * PITCONST * (1 << TIMER_SHIFT)));
         }
         else
         {
